backend/requirements.txt
asgiref==3.3.1
autopep8==1.5.4
Django==3.1.4
django-cors-headers==3.5.0
djangorestframework==3.12.2
djangorestframework-jsonapi==4.0.0
gunicorn==20.0.4
inflection==0.5.1
psycopg2==2.8.6
pycodestyle==2.6.0
pytz==2020.4
sqlparse==0.4.1
toml==0.10.2
